outsource merge-patch logic to separate helper class
